api_db/requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
pydantic==1.8.2
SQLAlchemy==1.3.24
asyncpg==0.22.0
databases==0.4.3
PyMySQL==0.9.3

starlette==0.14.2
pytest==6.2.3
PyYAML==5.4.1
freezegun==1.1.0